chore(irify): 清理审计风格模块与无关 electron-builder 修改
将 Irify 专用 audit style 与 focus loop 逻辑移至 irifyAiCodeAuditStyle.ts；撤销本地 mac 打包对 electron-builder toolsets 的 workaround，保持与 master 一致。
