+++ app/renderer/src/main/src/pages/irifyAiCodeAudit/IrifyAiCodeAuditHistoryItem.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import classNames from 'classnames'
 import { OutlineAIIcon, OutlineQuestionmarkcircleIcon } from '@/assets/icon/outline'
 import { PublicAIAuditCodeIcon } from '@/routes/publicIcon'
-import { IrifyAiCodeAuditStyle } from '@/constants/focusMode'
+import { IrifyAiCodeAuditStyle } from './irifyAiCodeAuditStyle'
 import { useI18nNamespaces } from '@/i18n/useI18nNamespaces'
 import { YakRunnerHistoryProps } from './YakRunnerIrifyAiCodeAuditType'
 import styles from './IrifyAiCodeAuditHistoryItem.module.scss'
