+++ app/renderer/src/main/src/pages/irifyAiCodeAudit/IrifyAiCodeAuditOnboardingMask.tsx
@@ -5,7 +5,7 @@ import classNames from 'classnames'
 import { YakitButton } from '@/components/yakitUI/YakitButton/YakitButton'
 import { YakitInput } from '@/components/yakitUI/YakitInput/YakitInput'
 import { useHistoryAIReActChat } from '@/components/historyAIReActChat'
-import { IrifyAiCodeAuditStyle, isIrifyAuditStyleConfirmed, resolveIrifyFocusModeLoop } from '@/constants/focusMode'
+import { IrifyAiCodeAuditStyle, isIrifyAuditStyleConfirmed, resolveIrifyFocusModeLoop } from './irifyAiCodeAuditStyle'
 import { SystemInfo } from '@/constants/hardware'
 import { handleOpenFileSystemDialog } from '@/utils/fileSystemDialog'
 import { showYakitModal } from '@/components/yakitUI/YakitModal/YakitModalConfirm'
