+++ app/renderer/src/main/src/pages/irifyAiCodeAudit/IrifyAiCodeAuditPage.tsx
@@ -3,7 +3,7 @@ import { useMemoizedFn } from 'ahooks'
 import emiter from '@/utils/eventBus/eventBus'
 import { AuditCodePageInfoProps } from '@/store/pageInfo'
 import { irifyAiCodeAuditPageAiStore } from '@/pages/ai-agent/store/ChatDataStore'
-import { IrifyAiCodeAuditStyle, resolveIrifyFocusModeLoop, isIrifyAuditStyleConfirmed } from '@/constants/focusMode'
+import { IrifyAiCodeAuditStyle, resolveIrifyFocusModeLoop, isIrifyAuditStyleConfirmed } from './irifyAiCodeAuditStyle'
 import { HistoryAIReActChatProvider } from '@/components/historyAIReActChat'
 import { AIInputEvent } from '@/pages/ai-re-act/hooks/grpcApi'
 import { IrifyAiCodeAuditSidePanelLayout } from './IrifyAiCodeAuditSidePanelLayout'
